Update to work correctly with PA build 77443
The "New System" Button was removed in this build, so we now add an
"New Awesome System" button.

Adjusted the structure of the "planets" objects within the system to match
the new spec.

It is no longer necessary to generate a placeholder system, as the game's
default of loading a pre-generated system is fine.